requirements.txt
# ============================================
# DEPENDENCIAS - vkApp 2.0
# ============================================

# Framework web
Flask==3.0.0
Werkzeug==3.0.1

# Base de datos
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

# Autenticación
Flask-Login==0.6.3

# Variables de entorno
python-dotenv==1.0.0
